+++ ui-views/src/main/java/app/cash/paykit/ui/views/CashAppPayButton.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package app.cash.paykit.core.ui
+package app.cash.paykit.ui.views
 
 import android.content.Context
 import android.util.AttributeSet